+++ core/app_shared.go
@@ -34,8 +34,8 @@ func (a appShared) getSurvey(id string, orgID string, appID string) (*model.Surv
 	return a.app.storage.GetSurvey(id, orgID, appID)
 }
 
-func (a appShared) getSurveys(orgID string, appID string, userID *string, creatorID *string, surveyIDs []string, surveyTypes []string, calendarEventID string, limit *int, offset *int, filter *model.SurveyTimeFilter, public *bool, archived *bool, completed *bool, includeResponses *bool, sortByDateCreated *bool, unstrucProps map[string]interface{}) ([]model.Survey, error) {
-	return a.app.storage.GetSurveysWithResponses(orgID, appID, userID, creatorID, surveyIDs, surveyTypes, calendarEventID, limit, offset, filter, public, archived, completed, includeResponses, sortByDateCreated, unstrucProps)
+func (a appShared) getSurveys(orgID string, appID string, userID *string, creatorID *string, surveyIDs []string, surveyTypes []string, calendarEventID string, limit *int, offset *int, filter *model.SurveyTimeFilter, public *bool, archived *bool, completed *bool, includeResponses *bool, sortByDateCreated *bool, unstrucProps map[string]interface{}, query *string) ([]model.Survey, error) {
+	return a.app.storage.GetSurveysWithResponses(orgID, appID, userID, creatorID, surveyIDs, surveyTypes, calendarEventID, limit, offset, filter, public, archived, completed, includeResponses, sortByDateCreated, unstrucProps, query)
 }
 
 func (a appShared) createSurvey(survey model.Survey, externalIDs map[string]string) (*model.Survey, error) {